fix: extract spawnAsync, fix maxBuffer regression, fix CORS port
Address all three review comments from @Zhang-Henry:

1. Glob maxBuffer regression — spawnAsync was rejecting when stdout
   exceeded maxBuffer (child killed → non-zero exit → reject), causing
   Glob to silently return "(no matches)" on large repos.  The shared
   spawnAsync now resolves with truncated data when the child is killed
   by our own SIGTERM, while genuine non-zero exits still reject.
   Partial-line output is trimmed at the last newline to avoid phantom
   filenames.

2. CORS port bug — getFrontendPortSync() was called without the
   VITE_PORT fallback because REQUESTED_VITE_PORT was defined after
   the CORS middleware setup.  Now computes CORS_VITE_PORT early and
   passes it to getFrontendPortSync().

3. spawnAsync deduplication — extracted the copy-pasted function from
   openrouter.js, cli-chat.js, and user.js into a shared module at
   server/utils/spawnAsync.js.  The user.js copy now also gets
   maxBuffer/truncation support.
